core: Fix function prototype of mpp_Update_ACHN_notest_Wrapper()
The old version of this function only received 4 arguments and loaded the
fifth value from the third argument. However, this isn't the right way to
do it because mpp_Update_ACHN_Wrapper() needs to pass values that are
independent from the third argument. This change will allow
mpp_Update_ACHN_Wrapper() to depend on mpp_Update_ACHN_notest_Wrapper()
later.