core: Rearrange code in mpp_Process_Effect()
The code that reads the tick counter into r2 and compares it to zero to set
the Z flag was extracted from each individual function into the main one
to save a bit of memory.

This patch moves the code to each individual function and documents the
parameters actually used by them, which will make it easier to disassemble
individual functions rather than doing everything in one go.