ds: Send number of samples to the ARM7 at mmInit()
When the soundbank is in RAM it is possible for the ARM7 to read this
information from the first word of the header. However, if the soundbank
is in NitroFS, this word isn't available in RAM. The ARM9 reads that word
in mmInitDefault() temporarily, so it needs to be passed to the ARM7
manually.