core: ds: gba: Fix platform-specific sample structures
The information about the GBA and DS sample data struct has been taken
directly from mmutil (commit 92249adaabaa). File source/mas.c, function
Write_SampleData().

In practice, the old code meant that the first 4 bytes of the sample were
skipped in GBA mode. The C struct was correct but it was unused, only the
incorrect definitions for ASM code were used.

In DS mode it worked correctly because only the C structure was incorrect,
but that structure wasn't used. The definitions for ASM code were correct.