core: Fix DCT note map accesses
This bug was present in the assembly code as well. The note map isn't in
a fixed place in the mm_mas_instrument struct. It is placed after the
envelope data, which is of variable length.

This patch also cleans up the code that uses the note map in the function
mmChannelStartACHN(). It is correct, but hard to understand.