core: Clarify situation of mmSetPosition() and mmPosition()
Currently mmSetPosition() is an alias of mmPosition(), but only in the GBA
header.

This patch turns it around. mmSetPosition() is clearer, so it becomes the
main function name, and mmPosition() the alias.

Also, both names are now available in all headers (GBA and both CPUS of the
DS) for consistency.