ds9: Only update active status from ARM7 messages
The ARM9 code should be aware that any command sent to the ARM7 may take
a bit of time to be received. There may be situations in which the ARM7
refuses to start playing a song, and all the error checking is on the
ARM7, so it's better to use the values obtained from the ARM7 as the
only source of truth.