ds: Implement mmGetPositionRow() and mmGetPosition() on the ARM9
The ARM7 now sends the row and pattern every VBL so that the ARM9 knows
the position of the main module.

mmGetPositionTick() isn't as useful because the tick can be updated more
than once per frame, so they can be easily skipped. Patterns and rows
are much less likely to be skipped.