ds: Do all SFX handling on the ARM7
The current system causes situations in which the ARM9 thinks that a
SFX channel is permanently used, blocking it for future use. Normally,
the ARM9 marks a SFX channel as used before telling the ARM7 to play a
sound effect, and eventually ARM7 tells the ARM9 which SFX channels
become free as they end (the update message is sent once per frame).

However, it is possible to have this situation:

- The ARM9 selects SFX channel 3 for the sound effect. It marks it as
  used and sends a message to the ARM7 to play a sound effect in that
  channel.
- The ARM7 looks for a free mixer channel to play the sound effect, but
  they are all in use (by the module, other sound effects, etc).
- mmEffectEx() fails, so the ARM7 never starts playing the sound effect,
  which means it never ends. This means that there is never an update
  for the ARM9 telling it that channel 3 is free again.
- SFX channel 3 is now free on the ARM7, but the ARM9 thinks it's busy,
  and there is no way to fix the problem.
- This problem can happen multiple times until all channels are blocked
  on the ARM9 and no more sound effects can be played.

The fix in this PR makes the system a bit slower because the ARM9 needs
to wait for the ARM7 to return the SFX handle after the request to play
the sound. However, it makes the process safe.