core: docs: Add layer to MMCB_SONGMESSAGE song events
Previously, only the effect number was passed as a parameter. This makes
it hard to know if it was the main module or the jingle to trigger the
event. Normally it's the main module, but it could happen that the
jingle also triggers events, and there would be no way to tell them
apart.

This patch passes the layer in the top for bits of the parameter so that
backwards compatibility is maximized. It's unlikely that any user has
ever used this callback to receive events from a jingle, so this patch
won't change anything for them. Furthermore, events are rarely used in
general, so the probability of ever finding a compatibility issue is
almost 0.