core: Prevent crashes when playing songs with too many channels
When a module has more chanels than Maxmod mmReadPattern() would try to
write out of bounds of the module channel array. This was particularly
noticeable with jingles because they are limited to 4 channels, but it can
also happen with the main module if Maxmod is initialized with a low enough
number of channels (stored in mm_num_mch).

Now, detecting this issue will stop the playback of that song instead of
corrupting memory.