build: ds: Use -Os instead of -O2 in the ARM7 of the DS
The space we have for code is quite limited, so let's save a bit of space.
The hardware mixer doesn't require a lot of CPU, so it's fine if we try
to save space. The more CPU-intensive parts of the software mixer are
written in assembly so the optimization flag don't matter that much.