ds7: Set the max volume correctly
The original code didn't really work. It did an OR of the original volume
with the value 100, but that doesn't set the new value to 100, it sets
the old value OR the new volume. However, this didn't matter because the
volume was set to 127 by libnds. Because of this, the volume was always
set to 127 anyway.

This patch switches the value to 127 so that it actually uses the maximum
volume supported by the DS.