ds: Simplify reverb configure IPC command
The old command sends fewer words with the FIFO in some cases, but the
complexity of the code is pretty high, particularly on the ARM9. In
practice, most people are going to set up reverb once and leave it alone,
so optimizing the number of words sent through the FIFO isn't useful.

The best-case scenario is that the ARM9 only sends the flags to the ARM7
and no additional parameters (so the ARM9 only needs to send one word), but
this isn't a very realistic scenario. In most cases, the ARM9 will just
send all parameters to the ARM7 to setup the effect.

It's better to reduce the size of this function than to optimize a code
path that is almost impossible to be critical for the game.