core: Remove outdated comments
The mask of locked/unlocked channels is stored in mm_ch_mask.

There is nothing in the code of the DS mixer that prevents the top 16 bits
of mm_ch_mask from being used.

The comments removed by this commit seem to be outdated and written before
expanding the size of the parameter of the IPC commands.