ds7: Remove outdated signaling code for streams
This was required in the original version of Maxmod because it relied on
shared memory for the ARM7 to communicate with the ARM9 and let it know
if the stream had started or stopped.

Currently, ProcessNextMessage() in the ARM7 sends FIFO messages to the
ARM9 after mmStreamBegin() and mmStreamEnd(). The command
MSG_ARM7_STREAM_READY is the replacement for this.

Commit 0e6ec115bd8d ("ds: Simplify synchronization of mm_stream
commands") replaced the original synch code by FIFO messages and commit
72148f104e4a ("ds: Clean ARM7 to ARM9 FIFO messages") replaced the magic
numbers of the message IDs by defines.

The comments removed in this commit were probably forgotten when the
first commit was pushed.