ds7: Protect accesses to the command FIFO
The circular buffer that stores ARM9 commands in the ARM7 is critical. If
the code is interrupted while editing it, it's possible that the state of
the circular buffer will be inconsistent and the ARM7 will execute leftover
data in the buffer as if they were real commands. This will normally crash
the ARM7.

Thanks to @pyramidensurfer for the minimal test case that reproduced the
bug.