docs: Update copyright file
Dave mostly worked on the build and packaging system of the repository,
which isn't even present in the code after the migration to BlocksDS. He
didn't actually work on the library, so commit 92b6873c0e27 ("license:
Update list of copyrights in license file") shouldn't have added the
name to the list.

Considering that Dave didn't even add himself, I think it's clearer if
we remove this entry.