core: Implement mmPlayMAS() and deprecate mmPlayModule()
mmPlayModule() expects the user to pass the base address of the MAS file
plus the size of the header, which isn't even defined in the same header
as the function (it's defined in mm_mas.h).

It is being kept for compatibility purposes, but new code should use
mmPlayMAS() instead, which doesn't require the offset.